Add new feature: Bootloader activation out of main application (implements #8)

+++ Inc/crc.h
@@ -4,6 +4,6 @@
 /*
  * Calculate CRC32 of memory region
  */
-unsigned long crc32_calculate(const unsigned char *data, size_t len);
+uint32_t crc32_calculate(const unsigned char *data, size_t len);
 
 #endif
